refactor: separate metric collection
