+++ src/lib/metrics.ts
@@ -1,118 +1,12 @@
 import _ from 'lodash';
 import apmAgent from 'elastic-apm-node';
-import { monitorEventLoopDelay } from 'node:perf_hooks';
-import type { Knex } from 'knex';
-import type { Pool } from 'tarn';
 import type { Middleware } from 'koa';
-
-import { scopedLogger } from './logger.js';
-import { getMeasurementRedisClient, type RedisCluster } from './redis/measurement-client.js';
-import { USERS_TABLE } from './http/auth.js';
-import { dashboardClient } from './sql/client.js';
-import type { IoContext } from './server.js';
-
-const logger = scopedLogger('metrics');
-const eventLoopMonitorResolution = 10;
 
 export class MetricsAgent {
 	private readonly counters: Record<string, number> = {};
 	private readonly statsAvg: Record<string, number[]> = {};
 	private readonly statsMed: Record<string, number[]> = {};
 	private readonly statsMax: Record<string, number[]> = {};
-	private readonly asyncSeries: Record<string, number[]> = {};
-	private readonly timers: Record<string, NodeJS.Timeout> = {};
-
-	constructor (
-		private readonly fetchRawLocalSockets: IoContext['fetchRawLocalSockets'],
-		private readonly redis: RedisCluster,
-		private readonly sql: Knex,
-		private readonly fetchProbes: IoContext['fetchProbes'],
-	) {}
-
-	run (): void {
-		const loopMonitorP95 = monitorEventLoopDelay({ resolution: eventLoopMonitorResolution });
-		const loopMonitorP99 = monitorEventLoopDelay({ resolution: eventLoopMonitorResolution });
-		const loopMonitorMax = monitorEventLoopDelay({ resolution: eventLoopMonitorResolution });
-		loopMonitorP95.enable();
-		loopMonitorP99.enable();
-		loopMonitorMax.enable();
-
-		const toMs = (value: number) => Math.max(0, Number(value) / 1e6 - eventLoopMonitorResolution);
-
-		registerGuardedMetric('nodejs.eventloop.delay.p95.ms', () => {
-			const loopDelay = toMs(loopMonitorP95.percentile(95));
-			loopMonitorP95.reset();
-			return loopDelay;
-		});
-
-		registerGuardedMetric('nodejs.eventloop.delay.p99.ms', () => {
-			const loopDelay = toMs(loopMonitorP99.percentile(99));
-			loopMonitorP99.reset();
-			return loopDelay;
-		});
-
-		registerGuardedMetric('nodejs.eventloop.delay.max.ms', () => {
-			const loopDelay = toMs(loopMonitorMax.max);
-			loopMonitorMax.reset();
-			return loopDelay;
-		});
-
-		this.registerAsyncCollector(`gp.measurement.stored.count`, async () => {
-			const [ dbSize, awaitingSize ] = await Promise.all([
-				this.redis.reduceMasters<number>(async (accumulator, client) => accumulator + await client.dbSize(), 0),
-				this.redis.zCard('gp:in-progress-timeouts'),
-			]);
-
-			// running measurements use 3 keys
-			// finished measurements use 2 keys
-			// 1 global key tracks the in-progress measurements
-			return Math.round((dbSize - awaitingSize - 1) / 2);
-		}, 60 * 1000);
-
-		this.registerAsyncCollector(`gp.probe.count.local`, async () => {
-			return this.fetchRawLocalSockets().then(sockets => sockets.length);
-		}, 10 * 1000);
-
-		this.registerAsyncGroupCollector('global probe stats', async () => {
-			const probes = await this.fetchProbes();
-			const byContinent = _.groupBy(probes, probe => probe.location.continent);
-
-			const countByContinent = _(byContinent)
-				.mapKeys((_probes, continent) => `gp.probe.count.continent.${continent}`)
-				.mapValues(probes => probes.length)
-				.value();
-
-			return {
-				...countByContinent,
-				'gp.probe.count.adopted': probes.filter(probe => probe.owner).length,
-				'gp.probe.count.total': probes.length,
-			};
-		}, 10 * 1000);
-
-		this.registerAsyncGroupCollector(`user stats`, async () => {
-			const result = await this.sql(USERS_TABLE)
-				.count('user_type as c')
-				.groupBy('user_type')
-				.select<{ user_type: string; c: number }[]>([ 'user_type' ]);
-
-			const countByType = _(result)
-				.mapKeys(record => `gp.user.count.${record.user_type}`)
-				.mapValues(record => record.c)
-				.value();
-
-			return {
-				...countByType,
-				'gp.user.count.total': _.sum(Object.values(countByType)),
-			};
-		}, 60 * 1000);
-
-		const getDashboardPool = (): Pool<unknown> | undefined => {
-			return (this.sql.client as Knex.Client).pool as Pool<unknown> | undefined;
-		};
-
-		registerGuardedMetric('gp.db.dashboard.pool.used', () => getDashboardPool()?.numUsed());
-		registerGuardedMetric('gp.db.dashboard.pool.pending_acquires', () => getDashboardPool()?.numPendingAcquires());
-	}
 
 	recordMeasurementTime (type: string, time: number): void {
 		this.recordStats(`gp.measurement.time.${type}`, time);
@@ -180,51 +74,9 @@ export class MetricsAgent {
 			return value;
 		});
 	}
-
-	private recordAsyncDatapoint (name: string, value: number): void {
-		if (!this.asyncSeries[name]) {
-			this.registerAsyncSeries(name);
-		}
-
-		this.asyncSeries[name]!.push(value);
-	}
-
-	private registerAsyncSeries (name: string): void {
-		this.asyncSeries[name] = [];
-
-		registerGuardedMetric(name, () => {
-			const value = this.asyncSeries[name]!.at(-1);
-			this.asyncSeries[name] = [];
-			return value;
-		});
-	}
-
-	private registerAsyncCollector (name: string, callback: () => Promise<number>, interval: number): void {
-		this.timers[name] = setInterval(() => {
-			callback().then((value) => {
-				this.recordAsyncDatapoint(name, value);
-			}).catch((error) => {
-				logger.error(`Failed to collect an async metric "${name}"`, error);
-			});
-		}, interval);
-	}
-
-	private registerAsyncGroupCollector (groupName: string, callback: () => Promise<{ [k: string]: number }>, interval: number): void {
-		this.timers[groupName] = setInterval(() => {
-			callback().then((group) => {
-				Object.entries(group).forEach(([ key, value ]) => {
-					this.recordAsyncDatapoint(key, value);
-				});
-			}).catch((error) => {
-				logger.error(`Failed to collect an async metric group "${groupName}"`, error);
-			});
-		}, interval);
-	}
 }
 
-export const initMetricsAgent = (fetchRawLocalSockets: IoContext['fetchRawLocalSockets'], fetchProbes: IoContext['fetchProbes']) => {
-	return new MetricsAgent(fetchRawLocalSockets, getMeasurementRedisClient(), dashboardClient, fetchProbes);
-};
+export const metricsAgent = new MetricsAgent();
 
 export const captureSpan = <R>(name: string, fn: () => R): R => {
 	const span = apmAgent.startSpan(name, 'app', 'custom');
@@ -322,7 +174,7 @@ function median (values: number[]): number | undefined {
 	return (values[half - 1]! + values[half]!) / 2;
 }
 
-function registerGuardedMetric (name: string, callback: () => number | undefined): void {
+export function registerGuardedMetric (name: string, callback: () => number | undefined): void {
 	apmAgent.registerMetric(name, () => {
 		const value = callback();
 		return value || 0; // NaN/undefined => 0
